+++ sql/queries/feed_follows.sql
@@ -1,27 +1,26 @@
--- name: CreateFeedFollow :many
-WITH insert_feed_follows AS (
-  INSERT INTO feed_follows (id, created_at, updated_at, user_id, feed_id)
-  VALUES(
-    gen_random_uuid(),
-    NOW(),
-    NOW(),
-    $1,
-    $2
-  )
-  RETURNING *
+-- name: CreateFeedFollow :one
+WITH inserted_feed_follow AS (
+    INSERT INTO feed_follows (id, created_at, updated_at, user_id, feed_id)
+    VALUES ($1, $2, $3, $4, $5)
+    RETURNING *
 )
-SELECT iff.*, u.name as user_name, f.name as feed_name
-FROM insert_feed_follows iff
-JOIN feeds f ON f.id = iff.feed_id
-JOIN users u ON u.id = iff.user_id;
+SELECT
+    inserted_feed_follow.*,
+    feeds.name AS feed_name,
+    users.name AS user_name
+FROM inserted_feed_follow
+INNER JOIN feeds ON inserted_feed_follow.feed_id = feeds.id
+INNER JOIN users ON inserted_feed_follow.user_id = users.id;
+--
 
 -- name: GetFeedFollowsForUser :many
-SELECT ff.*, u.name AS user_name, f.name AS feed_name FROM feed_follows ff
-INNER JOIN feeds f ON f.id = ff.feed_id
-INNER JOIN users u ON u.id = ff.user_id
-WHERE u.name = $1;
+SELECT feed_follows.*, feeds.name AS feed_name, users.name AS user_name
+FROM feed_follows
+INNER JOIN feeds ON feed_follows.feed_id = feeds.id
+INNER JOIN users ON feed_follows.user_id = users.id
+WHERE feed_follows.user_id = $1;
+--
 
 -- name: DeleteFeedFollow :exec
-DELETE FROM feed_follows
-WHERE user_id = $1
-AND feed_id = $2;
+DELETE FROM feed_follows WHERE feed_id = $1 AND user_id = $2;
+--
